Implement jit_{prolog,leaf}_{f,d} to ensure proper stack alignment.
  These are counterparts of jit_prepare_{f,d} for jit functions, and like
jit_prolog, jit_prepare must be called, only once, even if with a zero
argument. Calling jit_prepare_{f,d} with an argument of zero is a noop.
  These macros are required to properly calculate stack offset of arguments
when they may be either in registers or stack, and, in the case of x86_64,
need to figure out if the stack is padded or not, as arguments uses 8 bytes,
but the stack is aligned at 16 bytes.

  Note that jit_allocai information must also be reset in jit_prolog, to
properly match the runtime, aligned stack layout.