Correct general ALU mul behavior in x86_64.
  General behavior in the sense that it should work as expected with
the exception that there is not a cleanly defined behavior of what
jit_mulr_i should result in 64 bits. Semantically, it should be:
	jit_mulr_i  r0, r1, r2	->	r0(i64) = r1(i32) * r2(i32)
	jit_mulr_ui r0, r1, r2	->	r0(u64) = r1(u32) * r2(u32)
but, usually one should use the _i or _ui expecting to have it to be
a cheaper operation, and only care about the bottom 32 bits, for
example, by using jit_bltr_i, that does exactly that.

  Also, about general behavior, the immediate operations now, properly
mean:
	jit_muli_i  r0, r1, im	->	r0(64) = r1(i64) * (i32)im
	jit_muli_ui r0, r1, im	->	r0(64) = r1(i64) * (u32)im

  So, in 64 bits, the top 32 bits after jit_alur_i and jit_alur_ui
should be undefined, and that is the approach being taken.