Correct alu div for _l and _ul modifiers in x86_64.
  This is another case where 32/64 bits register naming matters, and cause
some confusion. To still share the 32 bits code (and correct x86_64 div
with _i and _ui modifiers), it should be implemented a new, and already
proposed, schema for naming registers. That is, now it works in x86_64
because it switched to use _RXX names, but the usage of _EXX names in
the shared code is now broken at least for div_{i,ui}.