Optimize some common immediate values in x86_64.
  Also corrected a major problem in recent patches to better comment
what is going on, where it would not properly copy register contents
in operations like r0 = r1 + 0, r0 = r1 << 0 when r0 if r0 was not
the same register as r1, and corrected some comments that were wrong,
or not properly telling the logic.

  This change uses inc and dec opcodes for 1 and -1 add/sub immediate
operands.
  Multiplication was changed to negate if immediate is -1.

  The expanded and commented macros should make it easier to convert
portions of the code to use inline functions.