Correct wrong macro expansion of shift operations.
  The macros were expanded to better describe the logic, and make it
easier to convert to inline functions as well as spot optimization
opportunities, but that should be left to be done when adding inline
functions, otherwise macros can become very hard to understand, or
at least indent and match the proper parenthesis.

  Interestingly, it was passing the tests in
http://code.google.com/p/exl/source/browse/trunk/check/lightning/alu_lsh.tst
and rsh.tst even with completely wrong 32 bit logic in x86_64, but the
problem was triggered after extending the test cases to cover all logic
paths, i.e. add 0 as operand to check register copy is done in code like
"r0 = r1 + 0;" and also added checks to ensure that "r1 = r0 alu r1;"
as well as the already existing  "r0 = r0 alu r1;" and "r0 = r1 alu r2;"
tests.