Correct hmul for 32 bit registers and immediates.
  The previous implementation probably would only work with a few
register combinations, but now should work on all combinations,
only possible issue could be "hmulr %r0 %r0 %r0", but the test
cases should be updated to test this, when applicable. Now it pass
the 32 bit hmul tests in:
http://code.google.com/p/exl/source/browse/trunk/check/lightning/alu_hmul.tst

  The patch logic was changed a bit to also include a comment, with
an almost verbatim copy of what an inline function implementing the
operation  would look like.