Implement/correct hmul{i,r}_{l,ul} in x86_64.
  The 32 bits implementation had some comments in the middle of the
macros removed, as excess information may actually make it harder
to understand, and the comment above better describe the logic.

  The 64 bits version is almost a cut&paste, adjusting opcodes, but
instead of saving registers in the stack if required, it uses the,
currently reserved %r12 (JIT_REXTMP), and also the, currently not
used, %r15, now aliased to JIT_REXTMP2. This hopefully should be
temporary, as it is desirable to have better control of free registers,
as well as make all of them visible/usable to the logic using lightning.