Implement add{c,x}{r,i}_{l,lu} and correct sub{c,x}i_{i,ui}.
  sub{c,x}i_ui cannot be safely defined as add{c,x}i_ui by changing sign
of the immediate value because of the "special" 0x80000000 value.

  Since the add/sub and set carry, followed by add/sub with carry
operation is only defined for unsigned types, the macro definitions
was changed to define those, and added the signed ones for possible
compatibility, but they could be under #ifndef _ASM_SAFETY or some
other consistency check.

  Again, a comment describing the inline function was added before
the macro implementation.

  The repository should be rebased to remove intermediate patches,
historic is not important, only correctness, and provide a more clean
set of patches as candidate for upstream merge.