Correct problems in float to integer conversion
  x87 logic could (or always would) generate a bad jump, when truncating
a negative integer, and a new jit_patch_short_at macro was added, and
the same approach as the correction for eqr_{f,d} and ner_{f,d} for
x86_64 was used. That is, attempt to write more readable logic, for
easier maintenance.

  x86_64 was using 32 bit opcodes in some 64 bit conversions, what would
zero extend, and is now corrected.

  This pass the new cvt.tst test.

  Logic for testing the behavior of NaN/Inf was added to the test case,
but disabled by default. But, maybe it should be changed to not adc/sbb
in certain conditions.