Convert 32 bit conditional jump macros to inline functions.
  Also remove a weird workaround that was required on the initial
implementation.

  The code builds and works correctly with the lightning test tool,
and pass all tests in both, x86_32 and x86_64, but significant changes
still need to be done, even in the code already implemented.

  The "correction" to jit_mul_i (to work on x86_64) should also probably
be slightly changed back to the logic used elsewhere to handle opcodes
with 8/32 bit immediates, or, convert to the logic used in jit_mul, where
it does not "hide" what is being generated, and choose based on immediate
size.