Initial conversion of x87 code to use inline functions.
  Since jit_roundr_x_y is misnamed, a new jit_rintr_x_y was added,
providing the functionality of the previous jit_roundr_x_y, and a
new jit_roundr_x_y, currently only for x87 was added, and this new
construct properly rounds away from zero.

  An experimental, proof of concept, schema to choose functions
was added, but currently just a compile time macro, that allows
generating code for i386 or newer, i686 or newer, and also an
option to assume round to nearest is always in effect, otherwise,
a slower, but safe, version that adjusts the rounding mode was
also added.

  This pass all testing tool tests, as well as some new tests in
the new fcvt.tst file.