Actually detect cpu features instead of having hardcoded test macros
  This adds jit_get_cpu() to i386/funcs.h, that is declared as a
constructor (called automatically before main), but that would only
happen if using gcc, so, it has a static flag to ensure it is called
only once, if requires the user to call it.

  There is also now a global jit_flags, that is only used to store
information of rounding mode. This is mainly for experimental purposes
for now, as it should actually be a jit_local_state field, but
automatically initializing jit_local_state fields is still a todo.