Reorder and redefine ALU related macros.
  ALUBrm had an extra comma in one of the arguments to the _O_r_x macro,
but never caused an error as it is not called from anywhere.

  The new _s32 and _u32 macros were added, to trigger constants too
large (what was actually the main initial problem in x86_64), and,
because of this new change, changed some code to cast some unsigned int
values to int, for example, argument to jit_boaddi_i.

  The ALU macros were also changed to avoid using Os_xxx_sL, and instead
check in the definition if the immediate fits in a signed 8 bytes or
not, also adjusting "locally" the opcode. This avoids several repeated
evaluations of the macro to check if the immediate fits in 8 bytes, and
also makes the code easier to follow.

  Extra evaluation of _O_r_X argument was removed as _r_X does not use
the extra argument; it is #if 0'ed, in the logic to handle x86_64 rip
addressing mode. This can/should be added back later, if really
required.

  Also in this commit, changed jit_check8() to also "recognize" %ebx
as a register that can be addressed as 8 bits in i386. It indeed is,
and actually the macro was not updated when reworking in core.h because
it was useful to test some changes, and ended up forgotten...