Simplify more asm macros and better immediate range test
  Most i16/i32/i64 opcodes only have a different prefix, so, try to
share the logic when possible.
  Change MOVLir macro to assert the argument is unsigned in x86_64.
This was another major source of problems/confusions in earlier work,
as "movl" zero extends in 64 bits mode, while most other opcodes
sign extend.
  Correct TESTQir macro to no longer use JIT_REXTMP if the immediate
was too large; it was completely wrong, but also no longer needed,
as the upper layer code now properly checks range. Also, there is
no need to check if the immediate fits in signed 8 bits (probably
a leftover of a cut&paste), as there is no immediate 8 bit opcode;
with exception of x86_64 that uses a 32 bits immediate, the immediate
must match the width of the register.