Reorder asm*.h definitions to not need prototypes to functions
  There are still a few macros in asm.h that depends on wordsize, but
this drastically reduce the number of "#if __WORDSIZE == nn" macros.

  Biggest advantage of using inline functions is that it automatically
detects wrong syntax (there were some in byte/word macros), and by
following compiler messages, allow having functions defined before they
are used, and the major one is static type checking.

  Major issue is that this commit, if not any of the previous, is like
a flag of do not push it upstream :-) But it is ok, later maybe it can
be reworked a bit for more sequential diffs; but unlikely, as too much
work has been already put to get it to work on x86_64, have sse on
i386, etc.