Correct g++ warnings and errors generated by test tool
  Test tool is not c++ clean, so, this change corrects all warnings
and errors up to when compiling lightning.c.
  The addition of jit_gpr_t, jit_fpr_t, jit_idx_t, etc may actually
require some special attention, and changes, to have source compatibility
and not special cases for one architecture or another.

  Nevertheless, clean c++ compilation is very desirable, as well as
having it "forcing" proper use of enums. This also corrected a major
mistake that had not been triggered because the test cases never
allocated memory above 2Gb, but it had the jit_movi_l arguments
reversed, due to changing MOVIQir to jit_movi_l, as jit_movi_l checks
range, and can use 32 bit opcode so zero extend, if address is in the
low 32 bits...