Major pass in conversion to add a jit_state_t argument to functions
  This is the major drawback with the lightning schema of using cpp
macros for everything, and is required for source compatibility, but
actually, can become superior to the macro schema because it would
allow multiple threads with its own data structures.

  Also, inline functions actually compile a lot faster, lightning
test tool compiles now roughly 5+ times faster, and should generate
better code as most expressions are evaluated once, instead of
huge nesting of macro expansions, leaving another large amount of
duplicated expressions for the compiler.

  This patch is not enough to allow using "#define _jit my_jit_buffer"
as it is still required to add the extra argument to the asm*.h files;
this commit updates core*.h and fp*.h.