Rename helper sse macros for better self documentation
  This changes only the new macros added recently, and use the pattern
"i" and "f" for 32 bits, "l" and "d" 64 bits, for example sse_if_rr
means two registers, first argument is a gpr and second is a fpr, and
sse_fl_rm means first argument is fpr float32, and second is a pointer
to int64.
  The toplevel macros are being kept unchanged for the sake of compatibility
in the rare case some code elsewhere (not in lightning) relies on it.

  Extra work should be done to classify what version of sse is required
to generate that code, and either add workarounds with older versions,
or fallback to x87 if cannot do it with less than sse2, that is assumed
in the current code, not a problem for x86_64, but there are still mmx/sse
only computers.