Add the lightning test tool to the new contrib/check directory
  It is not built by default, and actually require lightning in the
include, as well as binutils devel instead of the code in opcode
subdirectory.
  At first it is added mainly for commodity, and to have easy access
to it. It is an extra, and more complete set of tests.
  If test.sh does not print any messages, it means it exhaustively
brute force tested almost all possible combinations, and almost all
code paths, but, while most tests check all register combinations,
it does not test all possible values, as well as it cannot test
detected cpu dependent code paths; some of these may require manually
setting the value, in a, currently undocumented and non portable way,
for example, the statement "jit_cpu.sse = jit_cpu.sse2 = 0;" to force
it to use x87 on i386 (or just modify the macro jit_sse2_p() ...).