Correct problems on calls to C function with multiple arguments
  The previous code would not work correctly with prototyped functions
with odd number of arguments, and the 32 bits logic for st(0) was not
correct, as it must pop the x87 stack, or it may fault due to an
fpu stack overflow.

  The corrected logic is a bit tricky, as it needs to "intercept" the
first jit_pusharg_t, first arg_t, and/or first jit_allocai call, to
fixup the stack alignment, for functions it *may* call.

  The complex logic that was required to correct the problems almost
calls to an api change, to have jit_prolog* and jit_prepare* defined as:
	jit_prolog(nint, nfloat, ndouble)
	jit_prepare(nint, nfloat, ndouble)

  jit_addi_{i,d} were defined to use them instead of the assembly macros,
and also, to have it check for add/sub 0.

  A minor bug was also corrected in the testing tool, because it would
not properly parse expressions like "$(1+2)" due to incorrect logic
of parsing doubles, and thinking the '+' was a syntax error.