Extend test case of calls to C functions to include callbacks
  The base test case was modified to subtract arguments, so that result
would be different if it did pass arguments in one order, and receive
in the reverse order. This also required to do a more complete test
on unsigned results.

  The new test case basically declares pointers to functions in the
C code, that are updated by the jit code to match the jit functions
created in c_call.tst, then, the jit code calls a C function, that
calls a the function pointer to the jit function, and returns what
the jit function calculated back to the first jit function.

  The test pass in i386 and x86_64 without the need of any patches
to the current code.