Rework arguments and stack management on i386
  This change adds a forward patch that is updated on jit_allocai and
jit_prepare*, if it needs more stack space. To not limit stack usage
to 127 bytes, it forces a emit of a "sub 12,%esp" on jit_prolog where
the immediate is 32 bit.
  When jit_prolog* or jit_allocai are called, if called, and if there
is not enough stack space, it will recalculate the appropriate value
to keep the stack aligned at 16 bytes and fit the alloca and arguments
space, and patch the %esp adjustment.
  This should be a cheap call, but, could be removed/reworked if a proper
jit_leaf is implemented. Either way, it should be far more efficient
for more complex functions, and possibly slightly worse for jit functions
that only calls one function.