Use jit_allocai to allocate stack space for float conversion
  jit_allocai is called on demand, so that it does not need to patch
anything neither preallocate space. On the first need, it allocates
8 bytes, 4 bytes aligned, relative to %ebp, and keep it until the next
call to jit_prolog.
  The code that was not converted is sse_round_x_y to avoid confusing
code until a similar approach is adapted to x86_64, for both, patch
stack padding and update of jit_allocai logic. Another code still
using push/pop are the {sse,x87}_safe_xyz functions, that should be
called only when not assuming round to nearest or when a computer older
than a Pentium 6 was detected.
  Push/pop to save general purpose registers should be done separately
to avoid confusion or misuse, but these also should usually only happen
on special conditions, and/or, it may actually be cheaper/faster to do
a push/pop; floats are a special case because there is no push/pop
instruction with a x87 or mmx register argument.