Update x86_64 to match i386 logic to keep stack aligned at 16 bytes
  The new logic is actually simpler, and allows sharing more code, like
the change to fp-sse.h, that now allocates float_offset at a multiple of
8, not 4, but otherwise, the logic is the same.

  Information about received arguments was removed, but an assertion
should be added, to ensure declared and retrieved arguments counter
matches.

  This now approach has a possible drawback, that is code jumping to
a label outside of a jit function body, or to another jit function
body, what may cause it to use bad stack offsets if it jumps to the
body of a function that used more stack slots.