Add a new jit_push_pop_p() predicate to avoid patching stack adjustment
  There may be a few special cases where it is not desirable to have
the code to patch stack alignment set in jit_prolog() based on number
of arguments of called functions, or, allocate 8 bytes, and 8 bytes
aligned offset relative from %ebp "on demand" for float conversion.

  The only cases are when generating code that jumps from function
body to function body.

  In this commit, is being preferred to keep the behavior of patching
the stack adjustment, as code that jumps from function to function
should be very uncommon, and the support is being "added back" because
it was a feature/side-effect that should have very rare uses, but that
was implicitly removed.