Add initial support for mips64 instructions
  The test system does not define __mips64__ neither __mips64, but uname
tells mips64. Some more tests should be done, but most likely, arguments
should be fetched from stack, or, need to find out the proper abi.

  The initial 64 bits code was also added in a non standard way, using
suffix _ll instead of _l, as __WORDSIZE is defined as 32, but registers
are expected to be 64 bits wide (need to learn a bit more, maybe the
system is just running in 32 bit mode or something...).