More work on lightning mips port
  This commit is also a checkpoint commit. It is only known to compile
for now, and is almost certainly broken. The problems are expected to
be corrected interactively once the test cases in contrib/check can
be run, as those test all lightning interfaces and register combinations
as well as using some magic/special values to trigger possible problems
with signed/unsigned computations.

  This change mostly adds more work on providing the full instruction
set in asm.h, following the pattern _NAME(alist) where name is the
opcode mnemonic and alist is the list or arguments.