Update mips port to have first set of testcases to run
  The test cases not listed in contrib/check/test.sh should work
correctly.
  The test cases that are know to pass are ldst.tst, ldst_xi.tst,
ldst_xr.tst, ldst_xr-clobber.tst, ldst_i.tst and movi.tst.
  Something should be done to provide proper support for building
function calls when there are double arguments. For now it just
requires that the double arguments be listed last, and grouped
together. This is because lightning requires building a stack frame
with arguments "pushed" from right to left, but on the mips ABI,
to not require keeping track of state, it is required to push from
left to right; what actually makes sense and is more desirable...