Rework function call arguments management to align doubles
  This limits the number of arguments a function can receive to 256,
for saving space but could be as high as 8192 (or more if not using
opcodes with immediate arguments).

  Now double arguments should work correctly and appear anywhere in
the argument list; actually, still need to test properly prototyped
functions, as there is the suspecting fact that gcc loads doubles
from stack using two instructions, but no jit code was tested so
far with prototyped functions.