Correct integer branch and add new flag for mips2 instructions
  _SEB macro was incorrect due to changes on the code to emit the opcode,
but when corrected, it just showed that the instruction is not implemented
in the test computer; same goes for _ROTR and _WSBH, required by the
byte swap interfaces, and that need to be implemented as the sample
computer runs in little endian mode.

  Now test.sh should work up to cvt.tst, but not pass all tests in
branch.tst. Branch on carry/overflow to be added in next commits.