Correct several issues with float arguments
  It should more closely match the proper abi now, but floats and doubles
appear to have a complex logic, and the current code is known to be broken
for the case:
	fn(float, float, double);
because the first floats use (and must due to hardware issues) two registers,
and use $f12 and $f14 in case of 4 floats in a row. The problem in the above
example is that the logic does not yet properly handle a0+a1 for float0,
a2+a3 for float1 and then the double in the stack.
  The problem is that floats maps as 8 bytes and 8 stack slots if passed
on registers, but 4 stack slots if passed on stack. Also, the logic probably
is wrong, as objdump does not show floats or doubles in $a0+$a1.