Add extra checks for supported instructions and better carry logic
  Carry now should be properly propagated after subsequent calls to
jit_add.x_ui and jit_sub.x_ui, but it is only tested with only one
jit_....x_ui call after a jit_....c_ui. Should be good for 64 bits
emulation on 32 bits, but for 128 bit integer emulation this may not
be good enough. Bignum implementation cannot be supported because
loads/stores from memory may affect carry (in x86 at least) if need
to compute an offset for example.