Correct problems with bo{add,sub}{r,i} on mips
  The logic added "later" to correct behavior of the implementation did
actually work the rule that the second register argument could be the
assembler temporary; to correct this, the number of "exported" registers
has been reduced by one because it is required another temporary. Hopefully
this should have another solution at some point, as having a reserved
register just for this case is too much overkill.

  The logic for unsigned values also was bogus because it was actually
updating the value twice.

  A new set of tests was added to trigger these cases early, as most
related test cases did not properly test the conditions corrected in
this commit.