Add jit generation for struct and union field dereference.
  Store not yet implemented.
  The logic for non basic types was changed to not dereference symbols
in emit_load() as that would require to have it returning a flag
telling about it, and possibly receiving an argument allowing it,
otherwise, it would have problems with parsing of "->". But, this
logic should be added later to generate proper "ldxi_x ..." instead
of "addi_p ...; ldi_p ..." (this is now in the big TODO list).