Basic tok_decl handling, make vectors not lvalues and correct local offset
  "tok_decl handling" must still be done slightly differently, as global
variables must be initialized to constants.
  Vectors are now checked in emit_set and emit_setop, to ensure that,
besides treating them as pointers, the base address is not allowed to
change.
  Local variables offsets were being incorrectly "patched". It still needs
a more complete rewrite, as the patch to also put first variables closer
to the base pointer actually may cause problems if double alignment is
different than pointer alignment, and double must be aligned; known
issue in TODO list...