Add simple branch logic and implement if/else constructs
  The logic uses a stack of states where it can "lift" jumps before
patching them, so, it should generate very few jump threading
opportunities, what is one of the basic optimizations that should
be added to the intermediate representation before feeding it to
lightning.
  Conditional andand and oror as well as comparison operations also
generate already the proper test and branch operation, so, it should
also generate very few opportunities to some basic optimizations
"merging" instructions.