Add a new vstack_reset() helper macro to cjit_emit.c
  The macro also does an assertion to ensure the vstack.offset value
did not fall below the given argument, as registers variables may be
allocated in the bottom of the value stack. In that case, ordering is
important, because it should have 3 registers free to expressions at
all time (actually, should work with 2).

  Also renamed cjit_node_jump to EJIT_NODE_ARG to better describe it
is a macro and also, because patches are not only for branches.