Redesign of abi interface to call functions and receive arguments
  The ejit_putarg_x macros were removed, but now the requirements are
an instruction to move unmodified float/double to int/long register,
only required by the mips abi, but there may be other uses. And another
instruction to generate the equivalent "ret" opcode ("jr $ra" on mips).

  This should help in reducing some overhead by only saving callee save
registers if they are actually modified, and more freedom to handle
register arguments, but logic still needs to be written, to use the
register associated with the variable in a value_t, but prevent it
from being modified when not applicable.

  This rework is still subject to changes, but is one of the better
alternatives to "finally" start implementing function calls.