More rework on argument/register handling
  The rework is being done incrementally, so this is a checkpoint
commit. Now, with exception of handling of float registers in integer
registers, only code in "thunder.c" knows about architecture specific
details. The enumerations for moving to/from unmodified int/float
still could be kept, as they are available in mips and sse, and
can be done, moving from/to processor/coprocessor in x87 (or elsewhere)
by using (stack?) memory.