Add initial logic to make function calls
  While the current approach should be better for a C jit compiler,
a significant amount of code should be reverted next, and instead
use unmodified lightning, and do not know about internals; this is
because the "cjit" code is just an "experiment", and is not intended
to invest too much time on it.
  The code should be changed (back) to copy arguments to stack or
registers, make a shadow copy of arguments to function calls, etc.