ARM: Make functional very simple test case working with doubles.
  There is a very nasty problem (again) with the fact that lightning
specification says that arguments are pushed right to left, and because
doubles apparently must be aligned at 8 bytes in stack and registers,
this makes calculation of offsets hard to implement. And, to make
things worse, the current approach of loading immediates may generate
one to four instructions based on value, what makes patching non
trivial (but a problem only if passing way too many 4,8,4,8 interleaved
sized arguments).

  The simple tests in arm.c show that it is required to pad doubles
to have printf printing the proper value, what is done by adding
an extra integer argument to the printf call. And the second test
uses the code in this commit, that "packs" and even split doubles
in register/stack arguments, but that does not work.