ARM: Add limited support for returning float/double values from jit
  The limited support is that it must be done as:
	jit_movr_T(JIT_FPRET, ...);	jit_ret();
or
	jit_movi_T(JIT_FPRET, ...);	jit_ret();
that is, must be the last instruction, because it otherwise will
overwrite JIT_R0 (and JIT_R1 for doubles).

  Also, to "fetch" the return from a function, need to call
	jit_retval_T(JIT_FPR(n));
just after the jit_finish or jit_calli, because the value is actually
in JIT_R0 (and JIT_R1 for doubles).

  These rules are what is expected for safety, and no tests in contrib/check
break because of it, but there may exist code that relies on JIT_FPRET
being an actual hardware register, mapped to JIT_FPR0.