ARM: Minor changes to pass make check
  Still failing on some of the more complete contrib/check tests, what
may require a small redesign of what has been done so far. Possibility
being considered is to not use r0-r3, but use r4-r6 as JIT_R(0-3)
and r7-r9 as JIT_V(0-3). Then it could keep JIT_FPR(0-2) in r0-r3,
avoiding a lot of loads and stores. Also, should relocate r11 (fp)
to match code generated by gcc, and hopefully get a proper gdb
backtrace.

  Failing tests are branch.tst (issues with unordered comparison,
that should be easy to correct), and stack.tst, that is crashing in
a memmove call, and may be related to the way stack frames are built.