ARM: Correct make check failure if armv5 not detected.
  The correction is the change to use the proper base register in
jit_ldxi_d.

  The code was also changed a bit, besides not really necessary, to
change the base register in calls like jit_addi_i(JIT_TMP, r0, i0)
to use JIT_FTMP as base pointer. Not a big issue, just that if i0
cannot be encoded as an immediate, it will use JIT_TMP as source
and destination in the addition, due to mov'ing the immediate to
JIT_TMP.