PPC: Handle r0 as register argument and rework macros to functions
  r0 has special meaning as argument to certain instructions, and,
besides lightning not directly making r0 available, it could be the
argument to a macro.
  Several macros were transformed to inline functions to make the
code clear and better match my work on a new lightning syntax based
jit. During conversion to inline functions, several "shortcuts" were
added, as in a function it is easier to write code to check argument
values and output more optimized assembly.