ARM: Add support for the hardfp abi.
  Only test case failing is varargs.tst, that requires some special logic
to properly handle, and the considered idea is to have some kind of call
to tell where the "..." ellipsis is in a function call, as at that point
it must switch the "softfp" abi. If a failure is generated, it should at
best print random error messages for float values, and at worst crash,
due to generating a printf call using the hardfp abi.
  All handled instructions were generated and disassembled to properly
implement "packed float arguments", and some mistakes on not exercised
code paths, or not tested macros were detected and corrected. Most
notably was the float to integer move that was not properly encoded but
working for even numbered registers, and also the vadd macro (not actually
used by lightning) for simd add.
  Corrected a bug not properly tested by existing test cases, where
stxi_{f,d} was using the incorrect register order.