ARM: Correct encoding of constants and missing else before if.
  Check of jump encoding was incorrect what was invalidating the
assertion before calling encode_thumb_{,cc_}jump, what could cause
it to add a literal 0xffffffff when patching a jump.
  Encoding of "wide" immediate for add/sub was incorrect but did
not trigger any failure in lightning test cases.
  Missing else before if was frequently comparing twice two registers,
that is, if it did pass the first predicate for a 2 byte instruction
it would do it, but then due to missing else, would also do a 4 byte
instruction to compare and set flags again.