ARM: Correct arm jit regressions after adding thumb jit
  Tested on armv7 on a distro built with --with-float=softfp chroot and
another with --with-float=hard. On the softfp chroot it was validated
the soft float code and pass all tests.

  In this commit it was also added (back) support for str.w for 12, that
was removed due to previously incorrectly encoding ldr and str, but also
in this commit, only used to push arguments and patch pushed arguments;
next commit should add back more 12 bit positive offset opcodes.
  Thumb and software float is now explicitly assert'ed, as support for
it was not added to fp-swf.h, so, thumb mode requires vfp.