ARM: Correct shift immediate and add extra thumb opcode generation
  Immediate shift was passing all contrib/check tests by accident due
to not testing shifts with larger values. Depending on the shift value,
it could have been encoded to something like "mov pc, reg, asr #3".

  The extra opcodes are "wide" loads and stores with 12 bit unsigned
offset and added shortcut to use 16 bit instructions in ad{c,x}r
when appropriate, as those instructions sets flags if not in an
IT block.