ARM: Add support for thumb2 and soft float.
  Tested and pass all tests, but only way to trigger it is by
commenting the test:
    if (!jit_cpu.vfp && jit_cpu.thumb)
        jit_cpu.thumb = 0;
Tested on a softfp chroot by setting jit_cpu.vfp to zero and
commenting the test above in lightning/arm/funcs.h.

  Thumb2 is now supported with soft float, but not automatically
detected (need to have thumb2), so, if no vfp is detected it will
fallback to use arm instruction set.

  Also in this commit, when testing the !NAN_TO_INT_IS_ZERO
condition in all setups, corrected THUMB2_MVN define to not
encode R15 on it, as it uses the same interface as 3 registers
instructions, passing R15 as an argument, what would trigger
an assertion due to define value setting a register field.